Refactored Admin Page and fixed being able to set Expiration Date before Schedule date
